requirements.txt
django
gunicorn
psycopg2-binary
pymemcache
requests
beautifulsoup4
html5lib
lxml
pymongo
pycryptodome
whoosh
pyjwt
pyyaml
yamlordereddictloader
calmjs.parse
pylint
pylint-django
black
isort
autoflake
